templates/python-django/pyproject.toml
[project]
name = "python-django"
version = "1.0.0"
description = "Django boilerplate optimized for agentic dev velocity"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.0,<6.0",
    "djangorestframework>=3.14,<4.0",
    "uvicorn>=0.25,<1.0",
    "python-dotenv>=1.0,<2.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.0,<9.0",
    "pytest-django>=4.7,<5.0",
    "pytest-cov>=4.1,<5.0",
    "factory-boy>=3.3,<4.0",
    "pytest-factoryboy>=2.7,<3.0",
    
    # Type checking
    "mypy>=1.8,<2.0",
    "django-stubs>=4.2,<5.0",
    "djangorestframework-stubs>=3.14,<4.0",
    
    # Linting
    "ruff>=0.3,<1.0",
    
    # Formatting
    "black>=24.0,<25.0",
    "isort>=5.13,<6.0",
    
    # Docker
    "docker>=7.0,<8.0",
]

[tool.uv]
package = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
testpaths = ["apps"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
]

[tool.coverage.run]
omit = [
    "*/tests/*",
    "*/migrations/*",
    "config/wsgi.py",
    "config/asgi.py",
    "manage.py",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["config*", "apps*"]
